Move some tests out of basic.test.tsx into separate files (#424)
