+++ tests/basic.test.tsx
@@ -1,21 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { act, cleanup, fireEvent, render } from '@testing-library/react'
-import create, {
-  State,
-  StateListener,
-  StateSelector,
-  PartialState,
-  EqualityChecker,
-  StateCreator,
-  SetState,
-  GetState,
-  Subscribe,
-  Destroy,
-  UseStore,
-  StoreApi,
-} from '../src/index'
-// import { devtools, redux } from '../src/middleware'
+import { act, fireEvent, render } from '@testing-library/react'
+import create, { StateSelector, EqualityChecker, SetState } from '../src/index'
 
 it('creates a store hook and api object', () => {
   let params
@@ -418,97 +404,6 @@ it('can set the store without merging', () => {
   expect(getState()).toEqual({ b: 2 })
 })
 
-it('can subscribe to the store', () => {
-  const initialState = { value: 1, other: 'a' }
-  const { setState, getState, subscribe } = create(() => initialState)
-  const listener = jest.fn()
-
-  // Should not be called if new state identity is the same
-  let unsub = subscribe(() => {
-    throw new Error('subscriber called when new state identity is the same')
-  })
-  setState(initialState)
-  unsub()
-
-  // Should be called if new state identity is different
-  unsub = subscribe((newState: { value: number; other: string } | null) => {
-    expect(newState && newState.value).toBe(1)
-  })
-  setState({ ...getState() })
-  unsub()
-
-  // Should not be called when state slice is the same
-  unsub = subscribe(
-    () => {
-      throw new Error('subscriber called when new state is the same')
-    },
-    (s) => s.value
-  )
-  setState({ other: 'b' })
-  unsub()
-
-  // Should be called when state slice changes
-  listener.mockReset()
-  unsub = subscribe(listener, (s) => s.value)
-  setState({ value: initialState.value + 1 })
-  unsub()
-  expect(listener).toHaveBeenCalledTimes(1)
-  expect(listener).toHaveBeenCalledWith(
-    initialState.value + 1,
-    initialState.value
-  )
-
-  // Should not be called when equality checker returns true
-  unsub = subscribe(
-    () => {
-      throw new Error('subscriber called when equality checker returned true')
-    },
-    undefined,
-    () => true
-  )
-  setState({ value: initialState.value + 2 })
-  unsub()
-
-  // Should be called when equality checker returns false
-  listener.mockReset()
-  unsub = subscribe(
-    listener,
-    (s) => s.value,
-    () => false
-  )
-  setState({ value: initialState.value + 2 })
-  unsub()
-  expect(listener).toHaveBeenCalledTimes(1)
-  expect(listener).toHaveBeenCalledWith(
-    initialState.value + 2,
-    initialState.value + 2
-  )
-
-  // Should keep consistent behavior with equality check
-  const isRoughEqual = (x: number, y: number) => Math.abs(x - y) < 1
-  setState({ value: 0 })
-  listener.mockReset()
-  const listener2 = jest.fn()
-  let prevValue = getState().value
-  unsub = subscribe((s) => {
-    if (isRoughEqual(prevValue, s.value)) {
-      // skip assuming values are equal
-      return
-    }
-    listener(s.value, prevValue)
-    prevValue = s.value
-  })
-  const unsub2 = subscribe(listener2, (s) => s.value, isRoughEqual)
-  setState({ value: 0.5 })
-  setState({ value: 1 })
-  unsub()
-  unsub2()
-  expect(listener).toHaveBeenCalledTimes(1)
-  expect(listener).toHaveBeenCalledWith(1, 0)
-  expect(listener2).toHaveBeenCalledTimes(1)
-  expect(listener2).toHaveBeenCalledWith(1, 0)
-})
-
 it('can destroy the store', () => {
   const { destroy, getState, setState, subscribe } = create(() => ({
     value: 1,
@@ -682,165 +577,4 @@ it('ensures a subscriber is not mistakenly overwritten', async () => {
 
   expect((await findAllByText('count1: 1')).length).toBe(2)
   expect((await findAllByText('count2: 1')).length).toBe(1)
-})
-
-it('can use exposed types', () => {
-  interface ExampleState {
-    num: number
-    numGet: () => number
-    numGetState: () => number
-    numSet: (v: number) => void
-    numSetState: (v: number) => void
-  }
-
-  const listener: StateListener<ExampleState> = (state) => {
-    if (state) {
-      const value = state.num * state.numGet() * state.numGetState()
-      state.numSet(value)
-      state.numSetState(value)
-    }
-  }
-  const selector: StateSelector<ExampleState, number> = (state) => state.num
-  const partial: PartialState<ExampleState, 'num' | 'numGet'> = {
-    num: 2,
-    numGet: () => 2,
-  }
-  const partialFn: PartialState<ExampleState, 'num' | 'numGet'> = (state) => ({
-    ...state,
-    num: 2,
-  })
-  const equalityFn: EqualityChecker<ExampleState> = (state, newState) =>
-    state !== newState
-
-  const storeApi = create<ExampleState>((set, get) => ({
-    num: 1,
-    numGet: () => get().num,
-    numGetState: () => {
-      // TypeScript can't get the type of storeApi when it trys to enforce the signature of numGetState.
-      // Need to explicitly state the type of storeApi.getState().num or storeApi type will be type 'any'.
-      const result: number = storeApi.getState().num
-      return result
-    },
-    numSet: (v) => {
-      set({ num: v })
-    },
-    numSetState: (v) => {
-      storeApi.setState({ num: v })
-    },
-  }))
-  const useStore = storeApi
-
-  const stateCreator: StateCreator<ExampleState> = (set, get) => ({
-    num: 1,
-    numGet: () => get().num,
-    numGetState: () => get().num,
-    numSet: (v) => {
-      set({ num: v })
-    },
-    numSetState: (v) => {
-      set({ num: v })
-    },
-  })
-
-  function checkAllTypes(
-    _getState: GetState<ExampleState>,
-    _partialState: PartialState<ExampleState, 'num' | 'numGet'>,
-    _setState: SetState<ExampleState>,
-    _state: State,
-    _stateListener: StateListener<ExampleState>,
-    _stateSelector: StateSelector<ExampleState, number>,
-    _storeApi: StoreApi<ExampleState>,
-    _subscribe: Subscribe<ExampleState>,
-    _destroy: Destroy,
-    _equalityFn: EqualityChecker<ExampleState>,
-    _stateCreator: StateCreator<ExampleState>,
-    _useStore: UseStore<ExampleState>
-  ) {
-    expect(true).toBeTruthy()
-  }
-
-  checkAllTypes(
-    storeApi.getState,
-    Math.random() > 0.5 ? partial : partialFn,
-    storeApi.setState,
-    storeApi.getState(),
-    listener,
-    selector,
-    storeApi,
-    storeApi.subscribe,
-    storeApi.destroy,
-    equalityFn,
-    stateCreator,
-    useStore
-  )
-})
-
-type AssertEqual<Type, Expected> = Type extends Expected
-  ? Expected extends Type
-    ? true
-    : never
-  : never
-
-it('should have correct (partial) types for setState', () => {
-  type Count = { count: number }
-
-  const store = create<Count>((set) => ({
-    count: 0,
-    // @ts-expect-error we shouldn't be able to set count to undefined
-    a: () => set(() => ({ count: undefined })),
-    // @ts-expect-error we shouldn't be able to set count to undefined
-    b: () => set({ count: undefined }),
-    c: () => set({ count: 1 }),
-  }))
-
-  const setState: AssertEqual<typeof store.setState, SetState<Count>> = true
-  expect(setState).toEqual(true)
-
-  // ok, should not error
-  store.setState({ count: 1 })
-  store.setState({})
-  store.setState(() => undefined)
-  store.setState((previous) => previous)
-
-  // @ts-expect-error type undefined is not assignable to type number
-  store.setState({ count: undefined })
-  // @ts-expect-error type undefined is not assignable to type number
-  store.setState((state) => ({ ...state, count: undefined }))
-})
-
-it('should allow for different partial keys to be returnable from setState', () => {
-  type State = { count: number; something: string }
-
-  const store = create<State>(() => ({
-    count: 0,
-    something: 'foo',
-  }))
-
-  const setState: AssertEqual<typeof store.setState, SetState<State>> = true
-  expect(setState).toEqual(true)
-
-  // ok, should not error
-  store.setState((previous) => {
-    if (previous.count === 0) {
-      return { count: 1 }
-    }
-    return { count: 0 }
-  })
-  store.setState<'count', 'something'>((previous) => {
-    if (previous.count === 0) {
-      return { count: 1 }
-    }
-    if (previous.count === 1) {
-      return previous
-    }
-    return { something: 'foo' }
-  })
-
-  // @ts-expect-error Type '{ something: boolean; count?: undefined; }' is not assignable to type 'State'.
-  store.setState<'count', 'something'>((previous) => {
-    if (previous.count === 0) {
-      return { count: 1 }
-    }
-    return { something: true }
-  })
 })
